Add files via upload
This is my excel project on Supply Chain. The data used in this project is not real data, it is solely for the purpose of practicing and to display my capabilities of using Excel to perform data analysis. The data was in csv format so I converted it into an excel worksheet. I then checked for duplicates and missing values and there were none. I converted the numerical values from general to number and started working on the pivot tables, My visualization is based off of my pivot tables.